rename(controller): Rename package path of developer, socialauth controller
